Add more OOP
